ASGARD-1020 Stop task should redirect to "show"
Canceling a task running on a cluster object used to redirect correctly
to the show screen of a cluster. Recently it has been redirecting to a
nonsensical list URL, filtered on the cluster name (not the app name).
The resulting list is usually empty, and the list screen is the wrong
place to redirect anyway.

This behavior probably broke during the Grails 2 upgrade but no one
noticed or reported it.