ASGARD-1075 Disable form submit on keyboard enter key
The default browser behavior that allows the Enter key to submit the
form is dangerous and usually unnecessary. This danger has been
uncovered by a user navigating a select2 mutli-select input rapidly
with the keyboard, trying to add many ELBs to an ASG rapidly. The
form submitted early because of the Enter key, resulting in an ASG
that unintentionally lacked security groups.

For certain whitelisted, one-field forms that don't result in backend
state changes, the Enter key can be allowed to submit the form. That
makes those forms easier to use. For example, any form that merely
does a search or a browser redirect.