ASGARD-1091 Groovy in Java 7 fix ServiceUnavailableException
Convert all custom Groovy exceptions to Java in order to work around
Java 6-7 problem.

Java 7 changed the number of arguments in the constructor of the
Exception class. For some reason, Groovy does a runtime check on the
number of arguments in all the constructors of a Groovy class's
ancestor classes to see if they matches the arguments in the
compile-time constructors. We build in Java 6 but we want to run in
Java 6 and Java 7, so the workaround for now is to code all exceptions
in Java.

See https://github.com/Netflix/asgard/issues/244
and http://blog.proxerd.pl/article/how-to-fix-incompatibleclasschangeerror-for-your-groovy-projects-running-on-jdk7