ASGARD-1135 Prevent Firefox usage
There is a critical bug in Firefox on complex pages with many selects
that have the same name attribute. All other browsers behave normally,
but Firefox sporadically fails to mark the initially selected options in a
select element. This can cause deployments to occur without security
groups by accident, which can cause major outages.

Until we have a robust solution in place to work around the Firefox bug,
we are forbidding Firefox use in Asgard, as a safety measure. Users can
still use Chrome or Safari.