Prevent wasteful, concurrent cache fills
There is no benefit to filling a cache with two threads at the same time.
There is also no point in queuing up cache fill events one after another.
However, a thread that wants a cache to be recently filled before doing
other actions should wait for the current fill process to complete.

Instead, choreograph a deterministic dance between threads waiting for
each other to set appropriate state. This requires a hook to enable
multithreaded unit tests.

Also, referencing the private variable "journal" from within an inline
anonymous closure does not work correctly when this method is called
on a subclass of CachedMap, such as the MultithreadedTestingCachedMap
subclass. The simplest solution is to use ordinary for loops instead of
"each". This can also reduce Groovy's permgen usage because the
unnecessary closures do not need wrapper classes.