ASGARD-1117 - Prevent double form submission
It's common for people to click a submit button twice by accident or
due to impatience. When that happens for a case like Create Next ASG,
a second task will be created and will fail because the ASG name is
already taken. These errors are distracting, useless noise for the
user, the task list, the log, and Asgard developers.