ASGARD-1153 - Rolling Push block device mappings
Just as a Red Black Push completely replaces all Block Device Mappings based on the standard configuration for the new instance type, the Rolling Push should do the same. There is no convenient way for Asgard users to create their own choice of block device mappings anyway.

Even in the case where a power user created a Launch Configuration with the command line and chose their own Block Device Mappings, there is no consistent, reliable, simple algorithm to determine which of those mappings to retain, which ones to delete, and which ones to replace with standard mappings. We're working around Amazon's lack of a proper instance types API, and their lack of automatic helpful behavior for launching m3 instance types.