Forked test mode causes an issue where the CachedMap does not seem to be able to be mocked.
When I run tests outside of forked mode, they all pass correctly.

However, when I am running it in forked mode, they failed with the following message:

Cannot create mock for class com.netflix.asgard.CachedMap. To solve this problem, put Objenesis 1.2 or higher on the class path (recommended), or supply constructor arguments (e.g. 'constructorArgs: [42]') that allow to construct an object of the mocked type.

Seems that disabling fork mode for testing grails is a sensible thing to do