ASGARD-1261 - Create monitoring bucket type of "none"
A new monitor bucket type on the application config which allows for a monitoring type of "none".

This will result in an empty string for the value of the monitor bucket to be inserted into the user data of a launch config. This is part of our migration to a model that generally does not use this deploy-time setting for monitoring decisions.