Regression Suite -------------
This is a regression suite that runs against a real amazon aws instance and ensures core functionality within asgard works.

It assumes that the proper credentials have been added to the .asgard directory.

This set of tests also assumes there is an SSH key called nflxoss-asgard in the us-east region.

The set of tests are not ran by default as this would mean charges incurred against an AWS account.

To run the tests, call

./grailsw -DregressionSuite=true test-app functional:

This set of tests should be able to run against Saucelabs provided the correct credentials are provided.

The command to do so is:
./grailsw -DregressionSuite=true -Dgeb.env=sauce -Dsaucelabs.username=$SAUCE_USER_NAME -Dsaucelabs.key=$SAUCE_API_KEY test-app functional:

The Geb Configuration included in these set of tests also allow the RegressionSpec to be ran as a set of jUnit tests in IntelliJ idea granted the System property RegressionSuite=true is set as a default in jUnit tests.