Added KeyManagementServiceCredentialsProvider
Useful for situations where you have a sophisticated key management system with a known DNS name, and the local Asgard machines are Asgard developer laptops or workstations that have local key store files for signing SSL requests to the key management server.

Might also be useful for having cloud-resident Asgard instances fetch account credentials for other cloud accounts.

Currently this is written to conform to the contract of Netflix’s Cryptex service for which there is not yet an open source reference implementation.