Made it possible to list all other servers sharing a logical cluster with the current Asgard instance.
This will allow any one Asgard instance to figure out the server names of all the other Asgard instances that should be sharing visible state, such as a task list merged from all the in-memory task lists of all the Asgard instances in the cluster.

The cluster can also be determined a system property like -DotherServers=localhost:8081,localhost:8082 for testing and -DotherServers=asgard102:80 for logical clustering in a data center.