ASGARD-1285 - Withering
Asgard server endpoint to start Asgard ASG self-destruct "withering" process after running task count reaches zero.

While some of Asgard's long running tasks are still performed with in-memory threads instead of SWF, we need part of our deployment process to be a safe final clean up process.
Administrator sends Asgard a REST call to start withering.
An in-memory process starts polling for the running task count to reach zero.
When the tasks are all done, the withering process checks that it is the only instance in its ASG. If so, the withering process performs a force delete on the disabled ASG that contains the withering Asgard instance.