ASGARD-1302 Publish explicit Eureka VIP address
This will make it possible for people to make calls to Asgard in the cloud using Ribbon and Eureka instead of relying solely on DNS and ELBs.

Unfortunately, the behavior of the constructors in parent classes in eureka client makes it difficult or impossible to unit test this change without first rewriting eureka client to be more test-friendly.