ASGARD-1314 - Enable use of AssumeRole for AWS credentials to control a different account
This should work locally on a laptop through the use of a key management service, and should work in the cloud based on Amazon's STSAssumeRoleSessionCredentialsProvider class, which we feed with strings from ConfigService.

Unit tests will follow in a later commit. We need to try this cloud functionality pronto, and then clean up any weirdness afterward.