Support creation of cross hostedzone Route53 alias records.
Enhances the api to allow creation of alias records where the target
could be in a different hosted zone and also supports setting the target
health check flag on the alias record.
The different hosted zone is quite common if we are creating an alias to
an Amazon Load Balancer.
Also, without the health check flag being set, the AWS api will throw an
exception when creating an alias to an ELB.
Finally, when creating an alias record, the api was failing because an
empty ResourceRecord was being passed to the create request.