Suggested fix for issue #334
In order to support automatically taking any block device mapping of an AMI into an ASG they can be taken from the image itself. Any default EBS volumes on the AMI will be mounted as usual. 

This solves the issue we've been having with asgard not mounting the ephemeral disks in the AMI