Display HTML5 notifications when a task is no longer running
Sometimes bringing up a set of machines can take a while, and the user
would like to know when the process is complete.  Currently, the user
has to check back with Asgard every so often.  This commit introduces
the use of HTML5 notifications (which are not used if unsupported by
the user's browser) to inform the user when the task has completed or
failed.