restore unit test
This test invokes a constructor in RollingPushOperation which invokes
Spring.autowire(this).  The Spring.autowire method expects there to be
an application in the Holders.  This Spring.autowire method is suspect
and should probably be eliminated.  For now, mixing in
GrailsUnitTestMixin to this test should configure the holder so things
behave.